+++ truffle/src/com.oracle.truffle.api.debug/src/com/oracle/truffle/api/debug/DebuggerSession.java
@@ -33,6 +33,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.Predicate;
 
 import com.oracle.truffle.api.Assumption;
 import com.oracle.truffle.api.CompilerAsserts;
@@ -169,6 +170,8 @@ public final class DebuggerSession implements Closeable {
     private volatile boolean suspendAll;
     private final StableBoolean stepping = new StableBoolean(false);
     private final StableBoolean ignoreLanguageContextInitialization = new StableBoolean(false);
+    private boolean includeInternal = false;
+    private Predicate<Source> sourceFilter;
     private final StableBoolean breakpointsActive = new StableBoolean(true);
 
     private final int sessionId;
@@ -182,7 +185,7 @@ public final class DebuggerSession implements Closeable {
         if (Debugger.TRACE) {
             trace("open with callback %s", callback);
         }
-        addBindings();
+        addBindings(includeInternal, sourceFilter);
     }
 
     private void trace(String msg, Object... parameters) {
@@ -216,6 +219,16 @@ public final class DebuggerSession implements Closeable {
      */
     public void setSteppingFilter(SuspensionFilter steppingFilter) {
         this.ignoreLanguageContextInitialization.set(steppingFilter.isIgnoreLanguageContextInitialization());
+        synchronized (this) {
+            boolean oldIncludeInternal = this.includeInternal;
+            this.includeInternal = steppingFilter.isInternalIncluded();
+            Predicate<Source> oldSourceFilter = this.sourceFilter;
+            this.sourceFilter = steppingFilter.getSourcePredicate();
+            if (oldIncludeInternal != this.includeInternal || oldSourceFilter != this.sourceFilter) {
+                removeBindings();
+                addBindings(this.includeInternal, this.sourceFilter);
+            }
+        }
     }
 
     /**
@@ -360,31 +373,42 @@ public final class DebuggerSession implements Closeable {
         stepping.set(needsStepping);
     }
 
-    private void addBindings() {
+    private void addBindings(boolean includeInternalCode, Predicate<Source> sFilter) {
         if (statementBinding == null) {
             // The order of registered instrumentations matters.
             // It's important to instrument root nodes first to intercept stack changes,
             // then instrument statements, and
             // call bindings need to be called after statements.
-            Builder builder = SourceSectionFilter.newBuilder().tagIs(RootTag.class);
-            this.rootBinding = debugger.getInstrumenter().attachFactory(builder.build(), new ExecutionEventNodeFactory() {
+            this.rootBinding = createBinding(RootTag.class, includeInternalCode, sFilter, new ExecutionEventNodeFactory() {
                 public ExecutionEventNode create(EventContext context) {
                     return new RootSteppingDepthNode();
                 }
             });
-            builder = SourceSectionFilter.newBuilder().tagIs(StatementTag.class);
-            this.statementBinding = debugger.getInstrumenter().attachFactory(builder.build(), new ExecutionEventNodeFactory() {
+            this.statementBinding = createBinding(StatementTag.class, includeInternalCode, sFilter, new ExecutionEventNodeFactory() {
                 public ExecutionEventNode create(EventContext context) {
                     return new StatementSteppingNode(context);
                 }
             });
-            builder = SourceSectionFilter.newBuilder().tagIs(CallTag.class);
-            this.callBinding = debugger.getInstrumenter().attachFactory(builder.build(), new ExecutionEventNodeFactory() {
+            this.callBinding = createBinding(CallTag.class, includeInternalCode, sFilter, new ExecutionEventNodeFactory() {
                 public ExecutionEventNode create(EventContext context) {
                     return new CallSteppingNode(context);
                 }
             });
         }
+    }
+
+    private EventBinding<? extends ExecutionEventNodeFactory> createBinding(Class<?> tag, boolean includeInternalCode, Predicate<Source> sFilter, ExecutionEventNodeFactory factory) {
+        Builder builder = SourceSectionFilter.newBuilder().tagIs(tag);
+        builder.includeInternal(includeInternalCode);
+        if (sFilter != null) {
+            builder.sourceIs(new SourceSectionFilter.SourcePredicate() {
+                @Override
+                public boolean test(Source source) {
+                    return sFilter.test(source);
+                }
+            });
+        }
+        return debugger.getInstrumenter().attachFactory(builder.build(), factory);
     }
 
     private void removeBindings() {
@@ -519,8 +543,10 @@ public final class DebuggerSession implements Closeable {
     @TruffleBoundary
     void notifyCallback(DebuggerNode source, MaterializedFrame frame, Object returnValue, BreakpointConditionFailure conditionFailure) {
         // SuspensionFilter:
-        if (source.isStepNode() && ignoreLanguageContextInitialization.get() && !source.getContext().isLanguageContextInitialized()) {
-            return;
+        if (source.isStepNode()) {
+            if (ignoreLanguageContextInitialization.get() && !source.getContext().isLanguageContextInitialized()) {
+                return;
+            }
         }
         Thread currentThread = Thread.currentThread();
         SuspendedEvent event = currentSuspendedEventMap.get(currentThread);